+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added `accessibilityHint` to LabelColumn
 
 ## [7.0.3] - 2022-12-16
 - Set `TextInputView` clear button tint colour to `grey1` from `grey2`